backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database and ORM
sqlmodel==0.0.14
psycopg[binary]==3.1.17

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Additional utilities
python-multipart==0.0.6
